:recycle: Refactor kit filter form
Convert to TS and simplify the searchable select
usage, making it easier to add type definitions.
